# src/main.rs
use std::collections::HashMap;
use std::path::{Path, PathBuf};

use models::ResponseBuilder;
use rocket::fs::NamedFile;

#[macro_use]
extern crate rocket;
#[macro_use]
extern crate diesel;

mod common;
mod error;
mod models;
#[rustfmt::skip]
mod schema;

use diesel::prelude::*;
use lazy_static::lazy_static;
use rocket::http::Status;
use rocket::response::Redirect;
use rocket::serde::json::Json;
use std::env::var;
use toml::value::Table;
/// Database connection
#[rocket_sync_db_pools::database("postgres_database")]
struct UsersDbConn(diesel::PgConnection);

// Thinking about endpoints
// GET api/v1/student/{id}, request information about a user, must contain AUTH header with their token.
// POST api/v1/student/login, to request an access (JWT) token, must attach login details in body
// DELETE api/v1/student/{id}, to remove a user
// GET api/v1/highscores, to request a list of all the high scores
// POST /api/va/highscore, to add a new score - must contain information about the user

// TODO General todos
// Move common DB requests (such as looking up a user) into a framework under common.rs to avoid duplicate code.
// Modify get requests to support 500 server-failure errors if the db is unable to be accessed.

lazy_static! {
    static ref JWT_SECRET: String = var("JWT_SECRET").unwrap();
    static ref JWT_EXPIRY_TIME_HOURS: usize =
        var("JWT_EXPIRY_TIME_HOURS").unwrap().parse().unwrap();
    static ref BROWSER_BASE_URL: String = var("BROWSER_BASE_URL").unwrap();
    static ref COSTUMES: HashMap<String, models::Costume> = {
        //Load data from file, and parse as toml
        let data = std::fs::read_to_string("./costume.toml").expect("Unable to find `./costume.toml`");
        let f = data.parse::<toml::Value>().expect("Unable to parse `./costume.toml`");

        //Load global settings and costume data
        //TODO
        let _global: &Table = f.get("global").expect("Unable to find [global] tag in `./costume.toml`").as_table().expect("Unable to find [global] tag in `./costume.toml`");

        let costumes: &Table = f.get("costume").expect("Unable to parse `./costume.toml`, no costumes provided!").as_table().expect("Unable to find [global] tag in `./costume.toml`");

        //Validate all global settings are present
        //TODO

        //Parse each costume into hashmap
        let mut map: HashMap<String, models::Costume> = HashMap::default();
        let keys: Vec<&String> = costumes.keys().into_iter().collect();
        for key in keys {
            let costume = costumes.get(key).expect(&format!("Unable to parse costume {} from `./costume.toml`, is it correctly formatted?", key)).as_table().expect(&format!("Unable to parse {} as table from `./costume.toml`", key));
            let name: String = costume.get("name").expect(&format!("Unable to parse name for costume {} from `./costume.toml`", key)).as_str().expect(&format!("Unable to parse name for costume {} from `./costume.toml`", key)).to_owned();
            let description: String = costume.get("description").expect(&format!("Unable to parse description for costume {} from `./costume.toml`", key)).as_str().expect(&format!("Unable to parse name for costume {} from `./costume.toml`", key)).to_owned();
            let price: usize = costume.get("price").expect(&format!("Unable to parse price for costume {} from `./costume.toml`", key)).as_integer().expect(&format!("Unable to parse description for costume {} from `./costume.toml`", key)) as usize;

            map.insert(name.clone(), models::Costume {
                name,
                description,
                price,
            });
        }

        return map;
    };
}

/// Return information about the student
#[get("/api/v1/student")]
async fn get_student(token: models::Claims, conn: UsersDbConn) -> models::Response {
    //Load the item from the db, if it exists
    use crate::schema::users::dsl::*;
    let r: Option<crate::models::User> = conn
        .run(move |c| {
            let r = users
                .filter(id.eq(token.sub))
                .limit(1)
                .load::<crate::models::User>(c);
            if let Ok(mut v) = r {
                if v.is_empty() {
                    return None;
                }
                return Some(v.remove(0));
            }
            return None;
        })
        .await;

    // Format and return
    if let Some(user) = r {
        return models::ResponseBuilder {
            data: user,
            status: Status::Ok,
        }
        .build();
    }
    models::ResponseBuilder {
        data: "User Not Found",
        status: Status::BadRequest,
    }
    .build()
}

/// Attempt to login as a student
#[post(
    "/api/v1/student/login",
    data = "<login_information>",
    format = "application/json"
)]
async fn login_student(
    conn: UsersDbConn,
    login_information: Json<models::UserCredentials>,
) -> models::Response {
    let login_information = login_information.into_inner();
    //Check if the user exists in the db
    use crate::schema::users::dsl::*;
    let name = login_information.usr.clone();
    let r: Option<crate::models::User> = conn
        .run(move |c| {
            let r = users
                .filter(usr.eq(name))
                .limit(1)
                .load::<crate::models::User>(c);
            if let Ok(mut v) = r {
                if v.is_empty() {
                    return None;
                }
                return Some(v.remove(0));
            }
            return None;
        })
        .await;

    if r.is_none() {
        return models::ResponseBuilder {
            data: "Incorrect Password or Username",
            status: Status::BadRequest,
        }
        .build();
    }
    let r = r.unwrap();
    //Check that their password hash matches
    let hash_valid = match common::compare_hashed_strings(login_information.pwd, r.pwd) {
        Ok(h) => h,
        Err(e) => {
            return models::ResponseBuilder {
                data: format!("Failed to compare hashes {}", e.to_string()),
                status: Status::InternalServerError,
            }
            .build()
        }
    };

    if !hash_valid {
        return models::ResponseBuilder {
            data: "Incorrect Password or Username",
            status: Status::BadRequest,
        }
        .build();
    }

    return models::ResponseBuilder {
        data: models::Claims::new_token(r.id),
        status: Status::Ok,
    }
    .build();
}

/// Create a new student
#[post(
    "/api/v1/student/create",
    data = "<new_user>",
    format = "application/json"
)]
async fn create_student(
    conn: UsersDbConn,
    new_user: Json<models::UserCredentials>,
) -> models::Response {
    //Check their password meets minimum requirements
    let new_user = new_user.into_inner();
    if new_user.pwd.len() < 8 {
        return models::ResponseBuilder {
            data: "Password Too Short",
            status: Status::BadRequest,
        }
        .build();
    }

    //Check that the username isnt't taken
    use crate::schema::users::dsl::*;
    let name = new_user.usr.clone();
    let r: Option<crate::models::User> = conn
        .run(move |c| {
            let r = users
                .filter(usr.eq(name))
                .limit(1)
                .load::<crate::models::User>(c);
            if let Ok(mut v) = r {
                if v.is_empty() {
                    return None;
                }
                return Some(v.remove(0));
            }
            return None;
        })
        .await;

    if r.is_some() {
        return models::ResponseBuilder {
            data: "Username Taken",
            status: Status::BadRequest,
        }
        .build();
    }

    //Hash password
    let hashed_password = match common::hash_string_with_salt(new_user.pwd) {
        Ok(p) => p,
        Err(e) => {
            return models::ResponseBuilder {
                data: format!("Unable to hash password {}", e.to_string()),
                status: Status::InternalServerError,
            }
            .build()
        }
    };

    let new_user = models::UserCredentials {
        usr: new_user.usr,
        pwd: hashed_password,
        costumes: vec!["default".into()],
    };

    //Save account in db
    use schema::users;
    let r: Result<models::User, diesel::result::Error> = conn
        .run(move |c| {
            diesel::insert_into(users::table)
                .values(new_user)
                .get_result(c)
        })
        .await;

    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to insert into server {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    return models::ResponseBuilder {
        data: models::Claims::new_token(r.unwrap().id),
        status: Status::Created,
    }
    .build();
}

#[delete("/api/v1/student")]
async fn delete_student(token: models::Claims, conn: UsersDbConn) -> models::Response {
    //Check the user exists
    use crate::schema::users::dsl::*;
    let usr_id = token.sub.clone();
    let r: Option<crate::models::User> = conn
        .run(move |c| {
            let r = users
                .filter(id.eq(usr_id))
                .limit(1)
                .load::<crate::models::User>(c);
            if let Ok(mut v) = r {
                if v.is_empty() {
                    return None;
                }
                return Some(v.remove(0));
            }
            return None;
        })
        .await;

    if r.is_none() {
        return models::ResponseBuilder {
            data: "User Not Found",
            status: Status::BadRequest,
        }
        .build();
    }

    //Delete student from db
    let r: Result<crate::models::User, diesel::result::Error> = conn
        .run(move |c| diesel::delete(users.filter(id.eq(token.sub))).get_result(c))
        .await;

    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Unable to delete user due to error {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    return models::ResponseBuilder {
        data: format!("Account {} deleted", r.unwrap().usr),
        status: Status::Ok,
    }
    .build();
}

#[get("/api/v1/scores?<offset>&<limit>&<usr>&<id>")]
async fn get_scores(
    conn: UsersDbConn,
    offset: Option<i64>,
    limit: Option<i64>,
    usr: Option<String>,
    mut id: Option<i32>,
) -> models::Response {
    //Set the defaults for these values, and ensure non-negative
    let offset: i64 = offset.unwrap_or(0).abs();
    let limit: i64 = limit.unwrap_or(10).abs();

    if id.is_none() && usr.is_some() {
        //Load the id of the user suggested
        use crate::schema::users::dsl::{users, usr as usr_struct};
        let r: Option<crate::models::User> = conn
            .run(move |c| {
                let r = users
                    .filter(usr_struct.eq(usr.unwrap()))
                    .limit(1)
                    .load::<crate::models::User>(c);
                if let Ok(mut v) = r {
                    if v.is_empty() {
                        return None;
                    }
                    return Some(v.remove(0));
                }
                return None;
            })
            .await;
        if let Some(found_user) = r {
            id = Some(found_user.id);
        } else {
            let data: Vec<()> = vec![];
            return models::ResponseBuilder {
                data,
                status: Status::Ok,
            }
            .build();
        }
    }

    use crate::schema::scores::dsl::usr_id as struct_id;
    let r: Result<Vec<models::Score>, diesel::result::Error> = conn
        .run(move |c| {
            let mut db_request = crate::schema::scores::table.into_boxed();
            if let Some(usr_id) = id {
                db_request = db_request.filter(struct_id.eq(usr_id));
            }
            db_request
                .limit(limit)
                .offset(offset)
                .load::<crate::models::Score>(c)
        })
        .await;

    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to query the server due to error {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    let r = r.unwrap();
    if r.is_empty() {
        let data: Vec<()> = vec![];
        return models::ResponseBuilder {
            data,
            status: Status::Ok,
        }
        .build();
    }
    models::ResponseBuilder {
        data: r,
        status: Status::Ok,
    }
    .build()
}

#[post("/api/v1/scores", data = "<new_score>", format = "application/json")]
async fn add_score(
    token: models::Claims,
    new_score: Json<models::NewScore>,
    conn: UsersDbConn,
) -> models::Response {
    let new_score = new_score.into_inner();
    //Assign the user id
    let new_score = models::InsertableScore {
        usr_id: token.sub,
        score: new_score.score,
    };

    use schema::scores;
    let r: Result<_, diesel::result::Error> = conn
        .run(move |c| {
            diesel::insert_into(scores::table)
                .values(new_score)
                .execute(c)
        })
        .await;

    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to insert into server {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    return models::ResponseBuilder {
        data: "",
        status: Status::Created,
    }
    .build();
}

#[get("/api/v1/student/costumes")]
async fn get_costumes(token: models::Claims, conn: UsersDbConn) -> models::Response {
    //Load the user requested
    let search_id = token.sub;
    use crate::schema::users::dsl::*;
    let r: Result<Option<crate::models::User>, diesel::result::Error> = conn
        .run(move |c| {
            let r = users
                .filter(id.eq(search_id))
                .limit(1)
                .load::<crate::models::User>(c);
            return match r {
                Ok(mut v) => {
                    if v.is_empty() {
                        return Ok(None);
                    }
                    Ok(Some(v.remove(0)))
                }
                Err(e) => Err(e),
            };
        })
        .await;
    
    //Check request is ok
    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to query the server due to error {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }
    let r = r.unwrap();

    //Check user exists
    if r.is_none() {
        return models::ResponseBuilder {
            data: "User not found in database!",
            status: Status::NotFound,
        }
        .build();
    }

    //Return value
    return models::ResponseBuilder {
        data: r.unwrap().costumes,
        status: Status::Ok,
    }
    .build();
}

#[post(
    "/api/v1/student/costumes",
    data = "<costume>",
    format = "application/json"
)]
async fn unlock_costume(
    token: models::Claims,
    conn: UsersDbConn,
    costume: Json<models::Costume>,
) -> models::Response {
    let costume = costume.into_inner();
    //Check requested costume exists
    if !COSTUMES.contains_key(&costume.name) {
        return models::ResponseBuilder {
            data: "Costume does not exist",
            status: Status::BadRequest,
        }.build()
    }

    //Load costume
    let costume = COSTUMES.get(&costume.name).unwrap();

    //Load all scores from this user, and tally them to get their total score
    let search_id = token.sub;
    use crate::schema::scores::dsl::{scores, usr_id};
    let r: Result<Vec<crate::models::Score>, diesel::result::Error> = conn
        .run(move |c| {
            scores
                .filter(usr_id.eq(search_id))
                .load::<crate::models::Score>(c)
        })
        .await;
    
    //Check request is ok
    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to query the server due to error {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    //Tally their score
    let mut score = 0;
    for s in r.unwrap() {
        score += s.num_stars;
    }

    //Validate that they have enough stars
    if score < costume.price as i32 {
        return models::ResponseBuilder {
            data: "Costume is too expensive",
            status: Status::BadRequest,
        }.build()
    }

    //Modify that user with their new costume!
    let r = conn.run(move |c| {
        //HACK currently diesel does not support this sort of array manipulation, but it will come eventually!
        let cmd = format!("UPDATE users SET costume = array_append(costumes, {}) WHERE id={} LIMIT 1 RETURNING *;", &costume.name, &token.sub);
        diesel::sql_query(&cmd).load::<crate::models::User>(c)
    }).await;

    if let Err(e) = r {
        return models::ResponseBuilder {
            data: format!("Failed to query the server due to error {}", e.to_string()),
            status: Status::InternalServerError,
        }
        .build();
    }

    let r = r.unwrap();
    if r.is_empty() {
        return models::ResponseBuilder {
            data: format!("Failed to query the server due to being unable to find user! Was it deleted while this query was running?"),
            status: Status::InternalServerError,
        }
        .build();
    }

    return models::ResponseBuilder {
        data: r.get(1),
        status: Status::Ok,
    }.build()
}

/// Serve docs about the api
#[get("/api/docs")]
async fn docs() -> NamedFile {
    NamedFile::open(Path::new("static/docs/static.html"))
        .await
        .ok()
        .unwrap()
}

/// Returns the current health status of the database
#[get("/api/health")]
fn health() -> models::Response {
    //TODO
    ResponseBuilder {
        data: "Online",
        status: Status::Ok,
    }
    .build()
}

/// Handle the serving of any static resources for various pages
/// SAFETY: Rocket has a neat implementation preventing a path from getting outside of /static - keeping our host safe.
#[get("/api/static/<file..>")]
async fn website_resource(file: PathBuf) -> Option<NamedFile> {
    NamedFile::open(Path::new("static/").join(file)).await.ok()
}

/// Handle any 404's
#[catch(404)]
async fn not_found() -> Redirect {
    Redirect::to("/notfound")
}

#[launch]
fn rocket() -> _ {
    //TODO force all lazy_statics to parse here!
    rocket::build()
        .register("/", catchers![not_found])
        .mount(
            "/",
            routes![
                docs,
                get_student,
                login_student,
                create_student,
                delete_student,
                get_scores,
                add_score,
                unlock_costume,
                get_costumes,
                website_resource,
                health,
            ],
        )
        .attach(UsersDbConn::fairing())
}

#[cfg(test)]
mod tests {
    use super::rocket;
    use rocket::http::Status;
    use rocket::local::blocking::Client;

    #[test]
    fn test_health() {
        let client = Client::tracked(rocket()).expect("valid rocket instance");
        let response = client.get(uri!("/api/health")).dispatch();

        assert_eq!(response.status(), Status::Ok);
        assert_eq!(
            response.into_string(),
            Some("{\"data\": \"Online\"}".into())
        );
    }
}